add relationship info
move relationship info out of seed/index.js because it should be here along with the model export. Previously there because sequelize eats errors, leading to false assumptions. There was an errant relation definition that caused all relations to not be used (created but sequelize would through error if trying to use one of the relations)